Exception handling - delete

+++ src/main/java/com/fakezindev/webservicesspringboot/services/UserService.java
@@ -2,8 +2,11 @@ package com.fakezindev.webservicesspringboot.services;
 
 import com.fakezindev.webservicesspringboot.entities.User;
 import com.fakezindev.webservicesspringboot.repositories.UserRepository;
+import com.fakezindev.webservicesspringboot.services.exceptions.DatabaseException;
 import com.fakezindev.webservicesspringboot.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -32,8 +35,19 @@ public class UserService {
         return userRepository.save(obj);
     }
 
-    public void delete(long id) {
-        userRepository.deleteById(id);
+    public void delete(Long id) {
+
+        if (!userRepository.existsById(id)) {
+
+            throw new ResourceNotFoundException(id);
+
+        }
+        try {
+            userRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
